refactor: rename the Tag file component
